renomeando funções

+++ include/directory.hpp
@@ -8,7 +8,7 @@
 #include <vector>
 
 // Cria um novo diretório vazio em um cluster livre
-bool create_directory(const std::string& dirname, int parent_cluster);
+bool create_directory(const std::string& dirname, uint16_t parent_cluster);
 
 // Lista todas as entradas de um diretório
 std::vector<DirEntry> list_directory(int cluster_index);
@@ -17,9 +17,9 @@ std::vector<DirEntry> list_directory(int cluster_index);
 std::optional<DirEntry> find_entry(const std::string& name, int cluster_index);
 
 // Adiciona uma nova entrada em um diretório
-bool add_entry_to_directory(const DirEntry& entry, int cluster_index);
+bool add_entry_to_directory(const DirEntry& entry, uint16_t cluster_index);
 
 // Remove uma entrada de um diretório por nome
-bool remove_entry_from_directory(const std::string& name, int cluster_index);
+bool remove_entry_from_directory(const std::string& name, uint16_t cluster_index);
 
 #endif // DIRECTORY_HPP
